+++ CHANGELOG.md
@@ -20,6 +20,7 @@ BUG FIXES:
     because it exposes a bug with Puppet path lookup on VMware.
   - Fix bug in VirtualBox provider where port forwarding just didn't work if
     you attempted to forward to a port under 1024. [GH-1421]
+  - Fix cross-device box adds for Windows. [GH-1424]
 
 ## 1.1.0 (March 14, 2013)